manage_users/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.5
psycopg2-binary>=2.9
gunicorn>=21.2
python-decouple>=3.8
drf-spectacular>=0.27
